Move SimpleList component to global shared components.
