+++ client/a8c-for-agencies/components/simple-list/style.scss
