+++ client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-card/index.tsx
@@ -18,10 +18,10 @@ import { Icon, external, check } from '@wordpress/icons';
 import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
 import { useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
+import SimpleList from 'calypso/a8c-for-agencies/components/simple-list';
 import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { APIProductFamilyProduct } from 'calypso/state/partner-portal/types';
-import SimpleList from '../../common/simple-list';
 import { MarketplaceTypeContext } from '../../context';
 import { getHostingLogo, getHostingPageUrl } from '../../lib/hosting';
 import useHostingDescription from '../hooks/use-hosting-description';
