+++ client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-list/index.tsx
@@ -4,9 +4,9 @@ import { SiteDetails } from '@automattic/data-stores';
 import { useTranslate } from 'i18n-calypso';
 import { useMemo, useState, useContext } from 'react';
 import MigrationOffer from 'calypso/a8c-for-agencies/components/a4a-migration-offer';
+import SimpleList from 'calypso/a8c-for-agencies/components/simple-list';
 import useProductsQuery from 'calypso/a8c-for-agencies/data/marketplace/use-products-query';
 import { APIProductFamily } from 'calypso/state/partner-portal/types';
-import SimpleList from '../../common/simple-list';
 import { MarketplaceTypeContext } from '../../context';
 import useProductAndPlans from '../../hooks/use-product-and-plans';
 import { getCheapestPlan, getWPCOMCreatorPlan } from '../../lib/hosting';
