+++ client/a8c-for-agencies/sections/marketplace/wpcom-overview/wpcom-card/index.tsx
@@ -3,8 +3,8 @@ import { Button, JetpackLogo } from '@automattic/components';
 import formatCurrency from '@automattic/format-currency';
 import { useTranslate } from 'i18n-calypso';
 import { useContext } from 'react';
+import SimpleList from 'calypso/a8c-for-agencies/components/simple-list';
 import { APIProductFamilyProduct } from 'calypso/state/partner-portal/types';
-import SimpleList from '../../common/simple-list';
 import { MarketplaceTypeContext } from '../../context';
 import useWPCOMPlanDescription from './hooks/use-wpcom-plan-description';
